+++ docs/reference/changelog.rst
@@ -15,14 +15,21 @@ Changelog
 
     For a complete list of commits, check out the `1.2.3`_ release on GitHub.
 
-6.1.0 (YYYY-MM-DD)
+6.1.0 (unreleased)
 ------------------
+
+Services
+========
+
+- The Lifecycle service now has a ``get_plugin_group()`` method.
 
 Application
 ===========
 
 - If a lifecycle command is run with ``--destructive-mode``, but without root, a warning
   will be emitted about potentially unexpected behavior.
+- Direct registration of plugins in the Application is deprecated, as it's mutually
+  incompatible with the registration of plugin groups.
 
 Pytest plugin
 =============
@@ -520,8 +527,8 @@ Services
 - Setting the arguments for a service using the service factory's ``set_kwargs`` is
   deprecated. Use ``update_kwargs`` instead.
 
-Testing
-=======
+Testing with pytest
+===================
 
 - Add a :doc:`pytest-plugin` with a fixture that enables production mode for the
   application if a test requires it.